v1/backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2


